+++ packages/core/src/types.ts
@@ -1,0 +1,15 @@
+import { StyleCollector, CSSProps } from '@trousers/utils';
+
+declare module 'react' {
+    interface DOMAttributes<T> {
+        css?: StyleCollector<any> | CSSProps;
+    }
+}
+
+declare global {
+    namespace JSX {
+        interface IntrinsicAttributes {
+            css?: StyleCollector<any> | CSSProps;
+        }
+    }
+}
